requirements.txt
# This is a list of packages required for the project
python
flask==2.0.1
pyserial==3.5